--- src/auth/auth.service.ts
@@ -1,93 +1,0 @@
-import { Injectable, UnauthorizedException } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
-import axios from 'axios';
-
-@Injectable()
-export class AuthService {
-  constructor(private readonly configService: ConfigService) {}
-
-  /**
-   * 구글 로그인 페이지의 url을 반환한다.
-   */
-  async getGoogleLoginUrl() {
-    const { clientId, redirectUri } = this.getGoogleClient();
-    const scope = encodeURIComponent('profile email'); // 허용된 스코프는 프로필과 이메일이다.
-    const responseType = 'code';
-    const accessType = 'offline';
-
-    return `https://accounts.google.com/o/oauth2/v2/auth?client_id=${clientId}&redirect_uri=${redirectUri}&scope=${scope}&response_type=${responseType}&access_type=${accessType}`;
-  }
-
-  /**
-   * 구글 로그인 결과를 검증하고 jwt를 발급한다.
-   * @param code 클라이언트의 로그인 성공 시 얻을수 있는 코드값.
-   */
-  async getGoogleAuthorization(code: string) {
-    try {
-      const { accessToken, refreshToken } =
-        await this.getGoogleAccessToken(code);
-      const { uid, email, nickname } =
-        await this.getGoogleUserInfo(accessToken);
-
-      /**
-       * @todo member 조회후 토큰 발급
-       */
-      return { uid, email, nickname, accessToken, refreshToken };
-    } catch (error) {
-      throw new UnauthorizedException();
-    }
-  }
-
-  private getGoogleClient() {
-    return {
-      clientId: this.configService.get<string>('GOOGLE_CLIENT_ID'),
-      clientSecret: this.configService.get<string>('GOOGLE_CLIENT_SECRET'),
-      redirectUri: this.configService.get<string>('GOOGLE_REDIRECT_URI'),
-    };
-  }
-
-  private async getGoogleAccessToken(code: string) {
-    const { clientId, clientSecret, redirectUri } = this.getGoogleClient();
-
-    const response = await axios.post<{
-      access_token: string;
-      refresh_token: string;
-      token_type: 'Bearer';
-      scope: string;
-      expires_in: number;
-    }>(`https://oauth2.googleapis.com/token`, {
-      code,
-      client_id: clientId,
-      client_secret: clientSecret,
-      redirect_uri: redirectUri,
-      grant_type: 'authorization_code',
-    });
-
-    return {
-      accessToken: response.data.access_token,
-      refreshToken: response.data.refresh_token,
-    };
-  }
-
-  private async getGoogleUserInfo(accessToken: string) {
-    const response = await axios.get<{
-      id: string;
-      email: string;
-      verified_email: boolean;
-      name: string;
-      given_name: string;
-      family_name: string;
-      picture: string;
-    }>('https://www.googleapis.com/oauth2/v2/userinfo', {
-      headers: {
-        Authorization: `Bearer ${accessToken}`,
-      },
-    });
-
-    return {
-      uid: response.data.id,
-      email: response.data.email,
-      nickname: response.data.name,
-    };
-  }
-}
