+++ src/modules/auth.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { JwtModule } from '@nestjs/jwt';
+import { AuthController } from 'src/controllers/auth.controller';
+import { AuthService } from 'src/services/auth.service';
+
+@Module({
+  imports: [JwtModule],
+  controllers: [AuthController],
+  providers: [AuthService],
+})
+export class AuthModule {}
