+++ test/unit/auth.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
+import { AuthService } from '../../src/services/auth.service';
+import { AuthController } from 'src/controllers/auth.controller';
 
 describe('AuthController', () => {
   let controller: AuthController;
